Move all dom updates from the scripts.js to the domUpdates.js
